Cart page
Now, we have a functional cart page. That have all the items that the user is adding from the products list, and showing: image
title
tags
description
rating
price that is updated with discount if they have
remove button
When user is done, a checkout button is there and will direct user to a checkout page.